Build - disable code signing in appveyor
https://ci.appveyor.com/project/costlockerbot/desktop/build/1.0.40/job/9gaioyh26d1ne1on
Error: Cannot find certificate Costlocker SE, all certs:

https://www.electron.build/code-signing
However, it is not possible to export the EV Certificate as it is bound to a physical USB dongle. Thus, you can’t export the certificate for signing code on a CI, such as AppVeyor.
If you are using an EV Certificate, you need to provide win.certificateSubjectName in your electron-builder configuration.