Notifications - set app id before browser window is opened
Notifications are working in NSIS, but task bar icon is always blue.
Portable app: changeable icon, but no notifications (even when app is manually added to Start menu)

https://www.electron.build/configuration/nsis#guid-vs-application-name
It is also important to set the Application User Model ID (AUMID) to the appId of the application, in order for notifications on Windows 8/8.1 to function and for Window 10 notifications to display the app icon within the notifications by default. The AUMID should be set within the Main process and before any BrowserWindows have been opened, it is normally the first piece of code executed: app.setAppUserModelId(appId)